removed groups array and renamed to singular group
